Move invokeai/backend/grounded_sam -> invokeai/backend/image_util/grounded_sam
