+++ invokeai/backend/image_util/grounded_sam/detection_result.py
